Make unit tests independents
This commit introduces a way to separate execution of a single unit tests.
It helps to debug a single test instead of having to run the full suite.

Some tests are still depending from other ones (the thread ones) but that should not be a problem. Or we can clean that up later.